+++ src/spyglass/utils/dj_mixin.py
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 from time import time
 from typing import Dict, List
 
@@ -11,6 +12,21 @@ from spyglass.utils.database_settings import SHARED_MODULES
 from spyglass.utils.dj_helper_fn import fetch_nwb
 from spyglass.utils.dj_merge_tables import RESERVED_PRIMARY_KEY as MERGE_PK
 from spyglass.utils.dj_merge_tables import Merge
+=======
+from functools import cached_property
+from time import time
+from typing import Dict, List, Union
+
+import datajoint as dj
+from datajoint.expression import QueryExpression
+from datajoint.logging import logger as dj_logger
+from datajoint.table import Table
+from datajoint.utils import get_master, user_choice
+
+from spyglass.utils.dj_chains import TableChain, TableChains
+from spyglass.utils.dj_helper_fn import fetch_nwb
+from spyglass.utils.dj_merge_tables import RESERVED_PRIMARY_KEY as MERGE_PK
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
 from spyglass.utils.logging import logger
 
 
@@ -27,6 +43,14 @@ class SpyglassMixin:
         Fetch NWBFile object from relevant table. Uses either a foreign key to
         a NWBFile table (including AnalysisNwbfile) or a _nwb_table attribute to
         determine which table to use.
+    delte_downstream_merge(restriction=None, dry_run=True, reload_cache=False)
+        Delete downstream merge table entries associated with restricton.
+        Requires caching of merge tables and links, which is slow on first call.
+        `restriction` can be set to a string to restrict the delete. `dry_run`
+        can be set to False to commit the delete. `reload_cache` can be set to
+        True to reload the merge cache.
+    ddm(*args, **kwargs)
+        Alias for delete_downstream_merge.
     cautious_delete(force_permission=False, *args, **kwargs)
         Check user permissions before deleting table rows. Permission is granted
         to users listed as admin in LabMember table or to users on a team with
@@ -47,11 +71,15 @@ class SpyglassMixin:
         Alias for delete_downstream_merge.
     """
 
-    _nwb_table_dict = {}  # Dict mapping NWBFile table to path attribute name.
     # _nwb_table = None # NWBFile table class, defined at the table level
+<<<<<<< HEAD
     _nwb_table_resolved = None  # NWBFiletable class, resolved here from above
     _delete_dependencies = []  # Session, LabMember, LabTeam, delay import
     # pks for delete permission check, assumed to be on field
+=======
+
+    # pks for delete permission check, assumed to be one field for each
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
     _session_pk = None  # Session primary key. Mixin is ambivalent to Session pk
     _member_pk = None  # LabMember primary key. Mixin ambivalent table structure
     _merge_table_cache = {}  # Cache of merge tables downstream of self
@@ -62,96 +90,107 @@ class SpyglassMixin:
 
     # ------------------------------- fetch_nwb -------------------------------
 
-    @property
-    def _table_dict(self):
-        """Dict mapping NWBFile table to path attribute name.
-
-        Used to delay import of NWBFile tables until needed, avoiding circular
-        imports.
-        """
-        if not self._nwb_table_dict:
-            from spyglass.common.common_nwbfile import (  # noqa F401
-                AnalysisNwbfile,
-                Nwbfile,
-            )
-
-            self._nwb_table_dict = {
-                AnalysisNwbfile: "analysis_file_abs_path",
-                Nwbfile: "nwb_file_abs_path",
-            }
-        return self._nwb_table_dict
-
-    @property
-    def _nwb_table_tuple(self):
+    @cached_property
+    def _nwb_table_tuple(self) -> tuple:
         """NWBFile table class.
 
         Used to determine fetch_nwb behavior. Also used in Merge.fetch_nwb.
-        Multiple copies for different purposes.
+        Implemented as a cached_property to avoid circular imports."""
+        from spyglass.common.common_nwbfile import (
+            AnalysisNwbfile,
+            Nwbfile,
+        )  # noqa F401
 
-        - _nwb_table may be user-set. Don't overwrite.
-        - _nwb_table_resolved is set here from either _nwb_table or definition.
-        - _nwb_table_tuple is used to cache result of _nwb_table_resolved and
-            return the appropriate path_attr from _table_dict above.
-        """
-        if not self._nwb_table_resolved:
-            from spyglass.common.common_nwbfile import (  # noqa F401
-                AnalysisNwbfile,
-                Nwbfile,
+        table_dict = {
+            AnalysisNwbfile: "analysis_file_abs_path",
+            Nwbfile: "nwb_file_abs_path",
+        }
+
+        resolved = getattr(self, "_nwb_table", None) or (
+            AnalysisNwbfile
+            if "-> AnalysisNwbfile" in self.definition
+            else Nwbfile if "-> Nwbfile" in self.definition else None
+        )
+
+        if not resolved:
+            raise NotImplementedError(
+                f"{self.__class__.__name__} does not have a "
+                "(Analysis)Nwbfile foreign key or _nwb_table attribute."
             )
 
-            if hasattr(self, "_nwb_table"):
-                self._nwb_table_resolved = self._nwb_table
-
-            if not hasattr(self, "_nwb_table"):
-                self._nwb_table_resolved = (
-                    AnalysisNwbfile
-                    if "-> AnalysisNwbfile" in self.definition
-                    else Nwbfile
-                    if "-> Nwbfile" in self.definition
-                    else None
-                )
-
-            if getattr(self, "_nwb_table_resolved", None) is None:
-                raise NotImplementedError(
-                    f"{self.__class__.__name__} does not have a "
-                    "(Analysis)Nwbfile foreign key or _nwb_table attribute."
-                )
-
         return (
-            self._nwb_table_resolved,
-            self._table_dict[self._nwb_table_resolved],
+            resolved,
+            table_dict[resolved],
         )
 
     def fetch_nwb(self, *attrs, **kwargs):
         """Fetch NWBFile object from relevant table.
 
-        Implementing class must have a foreign key to Nwbfile or
-        AnalysisNwbfile or a _nwb_table attribute.
-
-        A class that does not have with either '-> Nwbfile' or
-        '-> AnalysisNwbfile' in its definition can use a _nwb_table attribute to
-        specify which table to use.
+        Implementing class must have a foreign key reference to Nwbfile or
+        AnalysisNwbfile (i.e., "-> (Analysis)Nwbfile" in definition)
+        or a _nwb_table attribute. If both are present, the attribute takes
+        precedence.
         """
-        nwb_table, path_attr = self._nwb_table_tuple
+        return fetch_nwb(self, self._nwb_table_tuple, *attrs, **kwargs)
 
-        return fetch_nwb(self, (nwb_table, path_attr), *attrs, **kwargs)
+    # ------------------------ delete_downstream_merge ------------------------
 
-    # -------------------------------- delete ---------------------------------
+    @cached_property
+    def _merge_tables(self) -> Dict[str, dj.FreeTable]:
+        """Dict of merge tables downstream of self: {full_table_name: FreeTable}.
 
-    @property
-    def _delete_deps(self) -> list:
-        """List of tables required for delete permission check.
-
-        Used to delay import of tables until needed, avoiding circular imports.
+        Cache of items in parents of self.descendants(as_objects=True). Both
+        descendant and parent must have the reserved primary key 'merge_id'.
         """
-        if not self._delete_dependencies:
-            from spyglass.common import LabMember, LabTeam, Session  # noqa F401
 
-            self._delete_dependencies = [LabMember, LabTeam, Session]
-            self._session_pk = Session.primary_key[0]
-            self._member_pk = LabMember.primary_key[0]
-        return self._delete_dependencies
+        self.connection.dependencies.load()
+        merge_tables = {}
+        for desc in self.descendants(as_objects=True):
+            if MERGE_PK not in desc.heading.names or not (
+                master_name := get_master(desc.full_table_name)
+            ):
+                continue
+            master = dj.FreeTable(self.connection, master_name)
+            if MERGE_PK in master.heading.names:
+                merge_tables[master_name] = master
 
+        logger.info(
+            f"Building merge cache for {self.table_name}.\n\t"
+            + f"Found {len(merge_tables)} downstream merge tables"
+        )
+
+        return merge_tables
+
+    @cached_property
+    def _merge_chains(self) -> Dict[str, List[dj.FreeTable]]:
+        """Dict of chains to merges downstream of self
+
+        Format: {full_table_name: TableChains}.
+
+        For each merge table found in _merge_tables, find the path from self to
+        merge via merge parts. If the path is valid, add it to the dict. Cache
+        prevents need to recompute whenever delete_downstream_merge is called
+        with a new restriction. To recompute, add `reload_cache=True` to call.
+        """
+        merge_chains = {}
+        for name, merge_table in self._merge_tables.items():
+            chains = TableChains(self, merge_table, connection=self.connection)
+            if len(chains):
+                merge_chains[name] = chains
+        return merge_chains
+
+    def _commit_merge_deletes(
+        self, merge_join_dict: Dict[str, List[QueryExpression]], **kwargs
+    ) -> None:
+        """Commit merge deletes.
+
+        Parameters
+        ----------
+        merge_join_dict : Dict[str, List[QueryExpression]]
+            Dictionary of merge tables and their joins. Uses 'merge_id' primary
+            key to restrict delete.
+
+<<<<<<< HEAD
     @property
     def _test_mode(self) -> bool:
         """Return True if test mode is enabled."""
@@ -218,12 +257,76 @@ class SpyglassMixin:
             (table & keys).delete(**kwargs)
 
     def delete_downstream_merge(
+=======
+        Extracted for use in cautious_delete and delete_downstream_merge."""
+        for table_name, part_restr in merge_join_dict.items():
+            table = self._merge_tables[table_name]
+            keys = [part.fetch(MERGE_PK, as_dict=True) for part in part_restr]
+            (table & keys).delete(**kwargs)
+
+    def delete_downstream_merge(
         self,
         restriction: str = None,
         dry_run: bool = True,
         reload_cache: bool = False,
         disable_warning: bool = False,
         return_parts: bool = True,
+        **kwargs,
+    ) -> Union[List[QueryExpression], Dict[str, List[QueryExpression]]]:
+        """Delete downstream merge table entries associated with restricton.
+
+        Requires caching of merge tables and links, which is slow on first call.
+
+        Parameters
+        ----------
+        restriction : str, optional
+            Restriction to apply to merge tables. Default None. Will attempt to
+            use table restriction if None.
+        dry_run : bool, optional
+            If True, return list of merge part entries to be deleted. Default
+            True.
+        reload_cache : bool, optional
+            If True, reload merge cache. Default False.
+        disable_warning : bool, optional
+            If True, do not warn if no merge tables found. Default False.
+        return_parts : bool, optional
+            If True, return list of merge part entries to be deleted. Default
+            True. If False, return dictionary of merge tables and their joins.
+        **kwargs : Any
+            Passed to datajoint.table.Table.delete.
+        """
+        if reload_cache:
+            del self._merge_tables
+            del self._merge_chains
+
+        restriction = restriction or self.restriction or True
+
+        merge_join_dict = {}
+        for name, chain in self._merge_chains.items():
+            join = chain.join(restriction)
+            if join:
+                merge_join_dict[name] = join
+
+        if not merge_join_dict and not disable_warning:
+            logger.warning(
+                f"No merge tables found downstream of {self.full_table_name}."
+                + "\n\tIf this is unexpected, try running with `reload_cache`."
+            )
+
+        if dry_run:
+            return merge_join_dict.values() if return_parts else merge_join_dict
+
+        self._commit_merge_deletes(merge_join_dict, **kwargs)
+
+    def ddm(
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
+        self,
+        restriction: str = None,
+        dry_run: bool = True,
+        reload_cache: bool = False,
+        disable_warning: bool = False,
+        return_parts: bool = True,
+<<<<<<< HEAD
         **kwargs,
     ) -> List[dj.expression.QueryExpression]:
         """Delete downstream merge table entries associated with restricton.
@@ -291,6 +394,38 @@ class SpyglassMixin:
             *args,
             **kwargs,
         )
+=======
+        *args,
+        **kwargs,
+    ) -> Union[List[QueryExpression], Dict[str, List[QueryExpression]]]:
+        """Alias for delete_downstream_merge."""
+        return self.delete_downstream_merge(
+            restriction=restriction,
+            dry_run=dry_run,
+            reload_cache=reload_cache,
+            disable_warning=disable_warning,
+            return_parts=return_parts,
+            *args,
+            **kwargs,
+        )
+
+    # ---------------------------- cautious_delete ----------------------------
+
+    @cached_property
+    def _delete_deps(self) -> List[Table]:
+        """List of tables required for delete permission check.
+
+        LabMember, LabTeam, and Session are required for delete permission.
+
+        Used to delay import of tables until needed, avoiding circular imports.
+        Each of these tables inheits SpyglassMixin.
+        """
+        from spyglass.common import LabMember, LabTeam, Session  # noqa F401
+
+        self._session_pk = Session.primary_key[0]
+        self._member_pk = LabMember.primary_key[0]
+        return [LabMember, LabTeam, Session]
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
 
     def _get_exp_summary(self):
         """Get summary of experimenters for session(s), including NULL.
@@ -317,6 +452,7 @@ class SpyglassMixin:
 
         return exp_missing + exp_present
 
+<<<<<<< HEAD
     @property
     def _session_connection(self) -> dj.expression.QueryExpression:
         """Path from Session table to self.
@@ -329,9 +465,25 @@ class SpyglassMixin:
                 connection if connection.has_link else False
             )
         return self._session_connection_cache
+=======
+    @cached_property
+    def _session_connection(self) -> Union[TableChain, bool]:
+        """Path from Session table to self. False if no connection found."""
+        connection = TableChain(parent=self._delete_deps[-1], child=self)
+        return connection if connection.has_link else False
+
+    @cached_property
+    def _test_mode(self) -> bool:
+        """Return True if in test mode.
+
+        Avoids circular import. Prevents prompt on delete."""
+        from spyglass.settings import test_mode
+
+        return test_mode
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
 
     def _check_delete_permission(self) -> None:
-        """Check user name against lab team assoc. w/ self * Session.
+        """Check user name against lab team assoc. w/ self -> Session.
 
         Returns
         -------
@@ -384,6 +536,7 @@ class SpyglassMixin:
                 )
         logger.info(f"Queueing delete for session(s):\n{sess_summary}")
 
+<<<<<<< HEAD
     @property
     def _usage_table(self):
         """Temporary inclusion for usage tracking."""
@@ -392,6 +545,14 @@ class SpyglassMixin:
 
             self._usage_table_cache = CautiousDelete
         return self._usage_table_cache
+=======
+    @cached_property
+    def _usage_table(self):
+        """Temporary inclusion for usage tracking."""
+        from spyglass.common.common_usage import CautiousDelete
+
+        return CautiousDelete
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
 
     def _log_use(self, start, merge_deletes=None):
         """Log use of cautious_delete."""
@@ -405,7 +566,10 @@ class SpyglassMixin:
             )
         )
 
+<<<<<<< HEAD
     # Rename to `delete` when we're ready to use it
+=======
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5
     # TODO: Intercept datajoint delete confirmation prompt for merge deletes
     def cautious_delete(self, force_permission: bool = False, *args, **kwargs):
         """Delete table rows after checking user permission.
@@ -463,6 +627,7 @@ class SpyglassMixin:
         """Alias for cautious_delete."""
         self.cautious_delete(*args, **kwargs)
 
+<<<<<<< HEAD
 
 class TableChains:
     """Class for representing chains from parent to Merge table via parts."""
@@ -591,3 +756,8 @@ class TableChain:
         for table in self.objects[1:]:
             join = join * table
         return join if join else None
+=======
+    def delete(self, *args, **kwargs):
+        """Alias for cautious_delete, overwrites datajoint.table.Table.delete"""
+        self.cautious_delete(*args, **kwargs)
+>>>>>>> b42432f0884ef3e990e111c53102a64529598ae5